+++ app/Http/Controllers/System/PaymentTermsController.php
@@ -54,6 +54,7 @@ class PaymentTermsController extends Controller
 	 
 	public function destroy($id){
 		$paymentTerm= PaymentTerm::findOrFail($id);
+		$this->authorize('delete', $paymentTerm);
 		$paymentTerm->delete();
 	
 	}
