+++ app/Http/Controllers/System/ProductGroupsController.php
@@ -49,8 +49,9 @@ class ProductGroupsController extends Controller
 		}
 	}
 	 
-	protected function destroy($id){
+	protected function destroy( $id){
 		$productGroup= ProductGroup::findOrFail($id);
+		$this->authorize('delete', $productGroup);
 		$productGroup->delete();
 	
 	}
