+++ app/Http/Controllers/System/TaxesController.php
@@ -51,6 +51,7 @@ class TaxesController extends Controller
     
     public function destroy($id){
     	$tax= Tax::findOrFail($id);
+    	$this->authorize('delete', $tax);
     	$tax->delete();
     }
 }
